# Conflicts:
#	core/src/com/game/Board/Board.java

+++ core/src/com/game/Board/Area.java
@@ -66,7 +66,7 @@ public class Area extends AssetManager {
     }
     public void setName(String string){this.name  =string;}
     
-    public void drawTexture(SpriteBatch sb, int xReduc, int yReduc) {
+    public void drawTexture(SpriteBatch sb, float xReduc, float yReduc) {
     	sb.draw(texture, xPos*xReduc, yPos*yReduc, 
     			(float) area.getWidth()*xReduc, (float) area.getHeight()*yReduc);}
